widgets: animimg: Port to LVGL 9.3
repeat_count argument type change to uint32_t

+++ src/lvglpp/widgets/animimg/animimg.cpp
@@ -27,7 +27,7 @@ namespace lvgl::widgets {
         lv_animimg_set_duration(this->raw_ptr(), duration);
     }
 
-    void AnimatedImage::set_repeat_count(uint16_t count) {
+    void AnimatedImage::set_repeat_count(uint32_t count) {
         lv_animimg_set_repeat_count(this->raw_ptr(), count);
     }
 }
